update script to handle moving the last message in a list
Previously, the script would return an error when attempting to select the top message in a mailbox if the last message you moved was the last one in the list. Now, the script first counts the messages in the list and only attempts to select one if there is >0 messages in the index.